Link Api Functions
Updated all links... 

Im not sure if this is the best way to address other possible updates to this page.  

Based on the proto docs, this one appears to be missing:  
link_entries	Consumes three values, two of which are the addresses of entries, and one of which is a string that defines a relationship between them, called a tag. Later, lists of entries can be looked up by using get_links. Entries can only be looked up in the direction from the base, which is the first argument, to the target.

Wondering if someone could confirm if that is intentional. 

Also, sensing it would help to add laymen's terms descriptions to these.  
Is that appropriate?  Since its the 'book' and intended for training, seems helpful.