Linux install script: do not install `sudo` [fix]
Without `sudo` available, and without being run as root,
we can not install `sudo`,
so previous code was bogus.

Also, a script for installing a software should not
by itsself install `sudo`,
which by some is considered a security risk.

Now it uses `sudo` if available, and otherwise requires to be run by root.